diplom_django/requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
psycopg
requests==2.31.0
urllib3<2
docker
pandas
djangorestframework